Updated analytics to exist in routes file vs root
